autonama.api/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
duckdb>=1.3.1

# Task Queue & Caching
celery>=5.5.3
redis>=5.2.0
flower>=2.0.1

# Data Processing & Analysis
pandas>=2.2.0
numpy>=1.24.0,<2.0  # Pin to NumPy 1.x for compatibility
scipy>=1.14.0

# Trading & Technical Analysis
ccxt>=4.4.0
optuna>=4.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Authentication & Security
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12

# HTTP & WebSocket
httpx>=0.28.0
websockets>=13.1
aioredis>=2.0.1
requests>=2.31.0
stripe>=8.0.0

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2
typer>=0.15.0

# Logging & ELK Integration
python-json-logger>=2.0.7
structlog>=24.4.0

# Development & Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.10.0
isort>=5.13.0
flake8>=7.1.0
mypy>=1.13.0

# Monitoring
# prometheus-client>=0.21.0
